refactor(ui): cap-split store halves merge back into one file each (KEN-550)
settings-zoom.ts, marketplaces-reads.ts, audit-run.ts and lib/adopt-shared.ts
each had exactly one importer and read only alongside it. Each folds back into
that importer; the helpers it exported for the sake of the split
(zoomActions, ZoomSlice, catalogReads, ReadCaches, auditRunner, replaceView)
become module-private. SharedLink and sharedLinkOf stay exported from
adopt-all.ts, where unmanaged-items.tsx and the adopt-all test now import
them. No behavior changes; the reads' `Set` type alias is renamed SetReads so
it no longer shadows the global in a larger file.

RATCHET_RAISE=1: fragments of one concept merged back into one file. New
baseline rows at the merged sizes: settings.ts 438, marketplaces.ts 327,
audit.ts 292. adopt-all.ts lands at 103, under the cap, so it takes no row.

The KEN-496 changelog entry is cut to three lines: the guard's entry-length
rule landed after it and was refusing every commit on the repo.
