+++ tools/size-ratchet-baseline.tsv
@@ -128,3 +128,6 @@ skills/size-ratchet/scripts/size-ratchet	1026
 skills/worktree/scripts/worktree	4548
 skills/worktree/scripts/worktree-session-guard	1197
 skills/worktree/tests/worktree_session_guard.sh	1282
+ui/src/stores/audit.ts	292
+ui/src/stores/marketplaces.ts	327
+ui/src/stores/settings.ts	438
